map-layer: Implement continuous zoom
The map layer now accepts fractional zoom levels by scaling the entire
map layer with gtk_snapshot_scale().

+++ shumate/shumate-scale.c
@@ -91,7 +91,7 @@ shumate_scale_compute_length (ShumateScale *self,
                               gboolean     *out_is_small_unit)
 {
   ShumateMapSource *map_source;
-  int zoom_level;
+  double zoom_level;
   double lat, lon;
   float scale_width;
   float base;
